Move Strava-specifics to strava.py
