+++ python/strava.py
@@ -1,3 +1,5 @@
+"""This module handles Strava API specific classes and API-calls."""
+
 import time
 import requests
 import swagger_client
@@ -20,3 +22,17 @@ def activity_to_dict(
         "elevation": activity.total_elevation_gain,
         "route": polyline.codec.PolylineCodec().decode(activity.map.summary_polyline),
     }
+
+
+def refresh_user(refresh_token) -> dict:
+    """Requests updated access credentials from the Strava API."""
+    r = requests.post(
+        "https://www.strava.com/api/v3/oauth/token",
+        data={
+            "client_id": secrets.STRAVA_CLIENT_ID,
+            "client_secret": secrets.STRAVA_CLIENT_SECRET,
+            "refresh_token": refresh_token,
+            "grant_type": "refresh_token",
+        },
+    )
+    return r.json()
